pyproject.toml
[tool.poetry]
name = "genai-video-creator"
version = "0.1.0"
description = "Agentic AI Video Creator using FastAPI, LangChain, LangGraph, and MoviePy"
authors = ["sgutha88 <guthas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
moviepy = ">=2.2.1,<3.0.0"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
langgraph = "^0.5.4"
langchain-community = "^0.3.27"
gtts = "^2.5.4"
python-multipart = "^0.0.20"
langchain = "^0.3.27"
imageio-ffmpeg = "^0.6.0"
imageio = "^2.37.0"
langchain-openai = "^0.3.28"
python-dotenv = "^1.1.1"
pydub = "^0.25.1"
aiofiles = "^24.1.0"
openai = "^1.97.1"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
